Update Shutdown() method
First checks if there is a root object and, if so, pushes it onto a stack. Then, it repeatedly pops an object from the stack and visits its references. If an object has not been visited before, it is marked as visited, its references are added to the stack, and its visit is continued. After all the objects reachable from the root have been visited and marked, the algorithm frees all the objects that were not marked.